# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/store/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersReadStoreController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get one of MY orders (404 on any order not mine — no IDOR)",
        "tags": [
          "Store / Orders"
        ]
      }
    },
...
    "/admin/v1/orders": {
      "get": {
        "operationId": "OrdersAdminController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending_payment",
                "paid",
                "fulfilled",
                "shipped",
                "delivered",
                "completed",
                "cancelled",
                "refunded",
                "partially_refunded"
              ]
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List orders (offset pagination, status / customer facets)",
        "tags": [
          "Admin / Orders"
        ]
      }
    },
    "/admin/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersAdminController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get an order by id (with items + status history)",
        "tags": [
          "Admin / Orders"
        ]
      }
    },
    "/admin/v1/orders/{id}/transitions": {
      "post": {
        "operationId": "OrdersAdminController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Drive one legal status transition (422 on an illegal edge)",
        "tags": [
          "Admin / Orders"
        ]
      }
    },
...
    "/admin/v1/stats/summary": {
      "get": {
        "operationId": "StatsAdminController_summary",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Dashboard summary KPIs with delta vs previous equal-length window",
        "tags": [
          "Admin / Stats"
        ]
      }
    },
    "/admin/v1/stats/timeseries": {
      "get": {
        "operationId": "StatsAdminController_timeseries",
        "parameters": [
          {
            "name": "root",
            "required": false,
            "in": "query",
            "allOf": [
              {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  }
                },
                "required": [
                  "from",
                  "to"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "granularity": {
                    "default": "day",
                    "type": "string",
                    "enum": [
                      "day",
                      "week",
                      "month"
                    ]
                  }
                },
                "additionalProperties": false
              }
            ],
            "x-nestjs_zod-unwrap-root": true,
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Zero-filled revenue+order-count timeseries for a date range",
        "tags": [
          "Admin / Stats"
        ]
      }
    },
    "/admin/v1/stats/top-products": {
      "get": {
        "operationId": "StatsAdminController_topProducts",
        "parameters": [
          {
            "name": "root",
            "required": false,
            "in": "query",
            "allOf": [
              {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  }
                },
                "required": [
                  "from",
                  "to"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "limit": {
                    "default": 5,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 20
                  },
                  "by": {
                    "default": "revenue",
                    "type": "string",
                    "enum": [
                      "revenue",
                      "quantity"
                    ]
                  }
                },
                "additionalProperties": false
              }
            ],
            "x-nestjs_zod-unwrap-root": true,
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Top products by revenue or quantity for a date range",
        "tags": [
          "Admin / Stats"
        ]
      }
    },
...
    "/admin/v1/stats/customer-breakdown": {
      "get": {
        "operationId": "StatsAdminController_customerBreakdown",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "New-vs-returning customer split for a date range (guests excluded)",
        "tags": [
          "Admin / Stats"
        ]
      }
    },
    "/admin/v1/stats/status-breakdown": {
      "get": {
        "operationId": "StatsAdminController_statusBreakdown",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Order counts grouped by status (all 9 statuses) for a date range",
        "tags": [
          "Admin / Stats"
        ]
      }
    },
...
  },
  "info": {
    "title": "SovEcom API",
    "description": "The SovEcom headless ecommerce API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TransitionOrderDto": {
        "type": "object",
        "properties": {
          "to": {
            "type": "string",
            "enum": [
              "pending_payment",
              "paid",
              "fulfilled",
              "shipped",
              "delivered",
              "completed",
              "cancelled",
              "refunded",
              "partially_refunded"
            ],
            "x-nestjs_zod-parent-additional-properties": false
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "to"
        ]
      },
...
    }
  }
}